Add ability to run from cached sigmoid fits and/or NOTA points
