+++ runFull.cpp
@@ -16,6 +16,8 @@
 #include "process.h"
 #include "fit.h"
 #include "test.h"
+#include "saveResults.h"
+#include "NOTAPoints.h"
 
 std::unordered_map<std::string, std::vector<ClassMember> > createTrainDataset(
     std::unordered_map<std::string, std::vector<ClassMember> > kSets[], size_t fold) {
@@ -114,14 +116,92 @@ std::vector<ClassMember> createTestSet(std::unordered_map<std::string, std::vect
     return testDataset;
 }
 
-void runFull(char* argv[]) {
+void addSigmoidFit(double* bestFitData, const std::string& className, const std::string& functionName) {
+    FitResult classSigmoid;
+    classSigmoid.c.setlength(2);
+    classSigmoid.c[0] = bestFitData[0];
+    classSigmoid.c[1] = bestFitData[1];
+    classSigmoid.functionName = functionName;
+    classSigmoid.wrmsError = bestFitData[2];
+    MODEL_STATE.bestFit[className] = std::move(classSigmoid);
+}
+
+void setSigmoidFitsFromFile(const std::string& sigmoidFitsFilename) {
+    std::ifstream sigmoidFitsFile(sigmoidFitsFilename);
+    std::stringstream ss;
+    std::string line, field, className, functionName;
+    std::unordered_map<std::string, std::unordered_map<std::string, double[3]> > bestFits;
+    std::unordered_map<std::string, std::unordered_map<std::string, size_t> > bestFitCount;
+
+    // Check whether the sigmoid fit data is per fold or for the overall class
+    std::getline(sigmoidFitsFile, line);
+    bool includesFold = line.rfind("fold", 0) != std::string::npos;
+
+    // Read all sigmoid fit data
+    while (std::getline(sigmoidFitsFile, line)) {
+        ss.str("");
+        ss.clear();
+        ss << line;
+        if (includesFold) {
+            std::getline(ss, field, ',');
+        }
+        std::getline(ss, className, ',');
+        std::getline(ss, functionName, ',');
+
+        for (size_t i = 0; i < 3; ++i) {
+            std::getline(ss, field, ',');
+            bestFits[className][functionName][i] += std::stod(field);
+        }
+
+        ++bestFitCount[className][functionName];
+    }
+
+    sigmoidFitsFile.close();
+
+    for (const auto& pair : bestFitCount) {
+        if (includesFold) {
+            // If the sigmoid fit data is per fold, calculate the overall parameter values for the class
+            auto mostFrequentFit = std::max_element(pair.second.begin(), pair.second.end(),
+                [](const std::pair<std::string, size_t>& fit1, const std::pair<std::string, size_t>& fit2) {
+                    return fit1.second < fit2.second;
+                });
+            double* bestFitData = bestFits[pair.first][mostFrequentFit->first];
+
+            for (size_t i = 0; i < 3; ++i) {
+                bestFitData[i] /= mostFrequentFit->second;
+            }
+
+            addSigmoidFit(bestFitData, pair.first, mostFrequentFit->first);
+        }
+        else {
+            auto bestFit = bestFits[pair.first].begin();
+            addSigmoidFit(bestFit->second, pair.first, bestFit->first);
+        }        
+    }
+
+    if (includesFold) {
+        writeBestFitFunctionsToCSV(CACHE_PATHS.bestFitFunctionsFilepath);
+    }
+}
+
+void runFull(int argc, char* argv[]) {
     std::string weightScheme = argv[1];
     std::string pValueThreshold = argv[2];
     bool applyFeatureWeighting = std::atoi(argv[3]);
     int numNeighborsChecked = std::atoi(argv[4]);
     int minSameClassCount = std::atoi(argv[5]);
-    std::string datasetFilename = argv[6];
-    std::string cacheDirectory = argv[7];
+
+    std::string sigmoidFitsFilename, NOTAPointsFilename, datasetFilename, cacheDirectory;
+    if (argc == 10) {
+        sigmoidFitsFilename = argv[6];
+        NOTAPointsFilename = argv[7];
+        datasetFilename = argv[8];
+        cacheDirectory = argv[9];
+    }
+    else {
+        datasetFilename = argv[6];
+        cacheDirectory = argv[7];
+    }
 
     std::unordered_map<std::string, std::vector<ClassMember> > dataset = readFormattedDataset(datasetFilename);
     std::unordered_map<std::string, std::vector<ClassMember> > kSets[K_FOLDS];
@@ -141,8 +221,14 @@ void runFull(char* argv[]) {
     MODEL_STATE.setDatasetSize();
     MODEL_STATE.setWeightExp(weightScheme);
     setPValueThreshold(pValueThreshold);
+    MODEL_STATE.preexistingBestfit = fileExists(sigmoidFitsFilename);
 
-    std::vector<ClassMember> NOTAPoints = createNOTAPoints(dataset);
+    std::vector<ClassMember> NOTAPoints = (fileExists(NOTAPointsFilename)) ? readNOTAPointsFromFile(NOTAPointsFilename)
+        : createNOTAPoints(dataset);
+
+    if (MODEL_STATE.preexistingBestfit) {
+        setSigmoidFitsFromFile(sigmoidFitsFilename);
+    }
 
     for (int fold = 0; fold < K_FOLDS; ++fold) {
         std::unordered_map<std::string, std::vector<ClassMember> > trainDataset;
@@ -180,8 +266,10 @@ void runFull(char* argv[]) {
         }
         std::cout << "Total in filteredDataset: " << totalFilteredPoints << "\n\n";
 
-        fitClasses(sorted_distances, fold);
-        
+        if (!MODEL_STATE.preexistingBestfit) {
+            fitClasses(sorted_distances, fold);
+        }
+
         std::vector<ClassMember> testDataset = createTestSet(kSets, NOTAPoints, fold);
 
         test(testDataset, fold);
