+++ runFull.h
@@ -1,6 +1,6 @@
 #ifndef RUN_FULL_H
 #define RUN_FULL_H
 
-void runFull(char* argv[]);
+void runFull(int argc, char* argv[]);
 
 #endif
